+++ tests/testthat/test_epochs.R
@@ -22,7 +22,8 @@ test_that("Test", {
                            sRates = c(100,100),
                            resample = 100,
                            epoch = 1,
-                           startTime = 0)
+                           startTime = 0,
+                           padding = 2)
 
   # Single vector
   epochs <- rsleep::epochs(signals = sig,